requirements.txt
requests
boto3
boto3-stubs
mypy-boto3-builder
line-bot-sdk